oauthutil: generalize bearer_token_command as shared OAuth option
Move bearer_token_command from an OneDrive-specific option to a shared
option in lib/oauthutil, so any OAuth backend gets it automatically.

When set, NewClientWithBaseClient skips the normal OAuth flow and
returns an HTTP client with a bearerTokenTransport that runs the
command to get a token, sets Authorization: Bearer on every request,
and re-runs the command on HTTP 401. An inert TokenSource is returned
so callers (e.g. NewRenew) always get a non-nil value.

Also fixes timeToExpiry to return ~95 years for nil tokens instead of
0, preventing the renewal timer from firing immediately on an inert
TokenSource.
